Fix edge case in u256 -> f64 conversion.
Handle special case when u256 number uses 53 or 54 bit.
The logic previously would fail because negative shifts
would occur for such numbers when calculating the sticky
bit.

The logic is now fixed by not simply taking the input
as significant for the 53 bit case and using a 0 bit
as the sticky bit for nmumber with 54 bits. For numbers
with 56 or more bits we can compute the sicky bit as
usual.