Update graph to correctly build Paths
Previously we did not take into account edge direction. Which would lead to nonsensical paths (swapping from same token to same token). Fixing this required saving the start token of each Path along with the edge indices, the remaining directions are then inferred by requiring the path to be properly connected.

This also required skipping involved addresses on KeySubsetIterator if they are not part of any path.