Don't run CI on beta and nightly toolchains
Builds got cancelled after 20 minutes total runtime.
I don't know the exact reason for the cancellation (it won't
tell me), and the docs suggest that the jobs are allowed to run
for much longer (6 hours), but maybe I am missing something
and the time cap of 20 minutes is actually in place.

As we don't aim to support features in nightly/beta, we don't
have an immediate need of trying them in CI.

I hope this change will allow the stable build to pass.

#time 10m