Use DatabaseRef instead of Database, use RefCell
Why: because Protosim protocol required it, and because Alan
wanted.

Pros: Now `SimulationEngine.simulate` takes an immutable reference
to self, which kind of looks nicer.

Cons: I had to do some magic with scopes to avoid simultaneous borrows.
Debugging this was much less pleasant than without RefCell,
because the errors show up only on runtime.
A `debug-cell` crate came in handy which prints the locations of borrows
in case a duplicated borrow is attempted.

--- don't change below this line ---
ROC-920 #time 2h 4m

#time 1h 5m