Allow passing big integers to Rust
- No need to go through String type to pass bug integers
  from Python to Rust
- SimulationParameters is now imported directly from Rust module,
  no need to define it on Python level
- Dropped "Py" prefix from Python-importable classes
- Added a dummy override to a python script to demonstrate
  that passing big integers works.

Background:

In Python, integers can be arbitrarily large. EVM storage slots
are of U256 type. So it should be possible to convert a Python integer
into an revm type as long as it fits into 256 bits.
However, to pass an integer from Python to Rust when using basic PyO3
mechanisms, we would need to go through a native u64 Rust type. This means
that a too large value would not fit there.

Fortunately PyO3 has a support for BigInt crate and offers conversion from
Python int to BigUint type. This type, in turn, can be converted to U256
needed by our simulation.

--- don't change below this line ---
ROC-1157 #time 48m