Fix Python demo
Previously we initialised the very same account
which we then attempted to call. This resulted in
a simulation error because we didn't set correct
contract code during account initialisation.

Now the initialisation/update/cache clear don't
have any effect on the simulation (they even happen
afterward), but the point is to just show how to
do it.

--- don't change below this line ---
ROC-1158 #time 38m