fix: Stale prices
The prices hashmaps had stale prices: the route ids used were not reliable at all. So we were not replacing the correct prices for the same route and these hashmaps were growing forever.

Use the id from RouteIdSubsetsByMembership in search_opportunities()

Make sure that everytime the graph is cleared and the routes calculated from scratch that the prices are also cleared and calculated from scratch. This way we don't have stale prices and we make sure that the prices reflect only active routes.

--- don't change below this line ---
ENG-3519 #time 1h 43m