fix: Return SimulationError from maybe_coerce_error
I've decided to add this back into the SimulationError class while just improving the docs to specify whether it is recoverable or not.

Reasoning:

I want to avoid the same we have in Python where we have very similarly named classes, though we have to keep in our own heads which ones are used where at which level, since they are slightly different.
I think reading the docs is enough - a separate class may simply be overkill.