fix: TEMPORARY
I simplified the overwrites logic... all the simulation overwrites are being done on the fly. we are only accumulating overwrites from previous simulations - in the case we did a simulation, changed the state and made a new one with some new block_lasting_overwrites set. These get merged with the on the fly overwrites.
This would remove the necessity for the state to be mutable in get_amount_out. However, there is still something wrong.. the tests don't pass without those extra hardcoded overwrites :'(