chore: Put default addresses back into create_engine.
This separation makes sense to Flo and me: the accounts which are always necessary for pre-compilation are set by default. Any other accounts which are specific to our state are handled in the state - such as the tokens and stateless accounts.

Sorry Diana if you don't like this we can move it back to set_engine.