fix: correctly insert balances updates
This commit fixes a bug where we were completely replacing the pool state.balances on updates. For >2 tokens pool it was leading to some balances being lost. Specifically, if the state.balances was holding balances for 3 tokens and the update was only affecting 2 of them, we were deleting the 3rd token's balance from the pool state. We are now just inserting the new balances but we keep the unchanged ones.