feat: split inserting static and dynamic parts of a new contract
The new versioning rules expect block-scoped inserts for dynamic/versioned data. Therefore it is important to group updates to each table into one gateway call each. If a gateway fn handles both a static value insert and a dynamic value insert this causes complications with subsequent updates applied to the dynamic values inserted. Therefore it is cleanest to maintain that each gateway fn insert/update only static OR dynamic values and not both.