fix: address a bug with `.returning` not returning on conflicts.
In SQL, when inserting data with ON CONFLICT DO NOTHING and RETURNING, it will ignore the conflict rows, so you can't expect this to return the full set of data for each row. We address this by first inserting and then reading, it comes with a small overhead but ensure we get the complete data.