refactor: update entry_point_tracing_result table to have a db internal id.
Before this commit we were using `entry_point_tracing_data_id` as primary key for this table. This works now but it's not future proof if we want to allow a one-to-many relation between the two tables (for example if we want to add some kind of versioning). This commit fixes this by adding a unique internal ID as primary key. The `entry_point_tracing_data_id` stays constrained to UNIQUE for now but this can be easily removed later if needed.