fix: correctly add default balances and code on account creations
This used to be done automatically in the previous version because `ref_into_account` uses `self.balance.clone().unwrap_or_default()`. Then these values used to be inserted together with the new contract. Now it's done in two steps, we still need to ensure that balance is set to default if `None`. This commit fixes this.