refactor: make `initialize_accounts` db inserts smaller
We got some cases in the testing suite where we where inserting many contracts with this. But the db transaction size became too big and it crashed. This commit address this by splitting inserts in many transactions. It makes it slightly slower but this is used for testing purposes only.