fix: Only pass storage values (not transient storage)
We would like to get the transient storage straight from the journaled state and not using an inspector (this is really expensive). However, to do that we need to access it before the transaction is over. This is only possible in revm 25.0.0 (this was released last week) and there isn't yet a compatible revm-inspectors neither foundry. We will wait for these releases before doing this feature (pass transient storage values from before swap to after swap simulations)

Took 2 hours 44 minutes