fix: correctly handle traced entrypoint upserts
Postgres doesn't allow for the same row to be updated twice in the same transaction. We need to handle this application side. Currently if an entrypoint is retriggered twice in a short timeframe, it can end up being upserted twice in the same db transaction. This commit makes sure we only try to upsert it once, for this we keep the latest updates if that happens.