fix: correctly fill the reorg buffer
This commit fixes a bug causing the block to not be found when calling `handle_finality`. This is because `handle_finality` was called before the new block was inserted. Also the block insertion and finality handling was happening only if `entrypoints_to_analyze` wasn't empty.

We are now always inserting a buffer layer for each block and always calling handle finality after processing everything. This should ensure that the latest finalized block is always found in the buffer.