fix: Don't return none for price if insufficient liquidity
Note:
This was necessary in order to normalize the TVL. I was checking for quite a large number of token prices, the last being the most common (WETH/BTC), and all were being filtered out by low liquidity when trying to find the price of the token for normalization (in this case BTC/WETH). My logic was that: a bad estimate was better than no estimate, at least for simple TVL filtering. This method should not be used for price calculation elsewhere.

Other fixes:
- Return error if setting up request builder fails.
- Track current timestamp, don't rely on response to do this