refactor: make extractor return a specific type
This commit improves how we deal with our types and communication with tycho-client.

The main objective here was to be able to detect any breaking change at compile time instead of runtime. We now deal with specific types to ensure serialization and deserialization are on the same Rust struct `dto::BlockChanges` and equivalent dto for `Command` and `Response`. We also removed the `NormalisedMessage` trait because it was not needed anymore.