feat: Specify action trait system
This specifies the action trait system and implements a new trait Swappable that is aimed to replace the ProtocolSim trait in the future.

This should serve as the first showcase that shows how to build a more highlevel trait from a more lowlevel trait implementation.

The action trait system is meant to be quickly extensible to other use cases this way tycho is ready to support any other onchain actions that do not fit into the standard Swappable interface.