fix: correctly get all values from cache when needed.
In some cases we need a full cache lookup. For example when getting retriggers or tracked slots we want to return all the elements and not only the last one.

This commit adds a `get_all` function on the `VersionedCache` to allow for this. We now use it instead of a simple `.get()` when needed inside DCI implementation