fix: correctly merge attributes when inserting "pausing".
This commit fixes a bug that was occuring when inserting "paused" attribute in DCI. We were replacing the current update in case of conflict. We are now correctly merging them.
Also, this same logic will be used by hook DCI, so this commit moves it outside of dci to avoid code duplication.