fix: prevent race condition in extractors/server startup sequence
Fix race condition where extractors could start receiving blocks before the RPC pending delta task was fully subscribed. This caused the pending delta to miss the first block, leading to corruption and Tycho crashes.

Changed startup to be sequential: server and pending delta tasks must be fully running before extractor tasks are spawned.

Unfortunately it's really hard to make a unit test for this. But this was tested locally against a real database.