fix: Solve issue with low balance tokens in pool manager
Sometimes when the pool manager doesn’t hold many AAA tokens, simulating on a hook might fail.

Since we are simulating only on the before/after swap code directly, this assumes that the user has already transferred the funds into the pool manager. If it is a common token this is usually fine (the pool manager already has a high balance of this token), however sometimes this is not the case and we need to mock the pool managers balance for this token

Took 6 minutes