fix: make `TracingResult` deserialization backward compatible
We added a custom deserialization that defaults offset to 0 when it's missing. This is fine because it's not used by tycho-simulation.

This is needed to smoothly update prod with the new release without having to sync with all clients. They can now update their tycho-client version before and it will automatically support both old and new release.

We can remove this once everything is more stable.