feat: keep stale protocol streams
Instead of crashing the BlockSynchronizer, we now keep it alive as long there is at least one healthy protocol.

To allow for more restrictive policies, we include the state of each protocol even if stale or ended within the output. Consumers of the stream can then decide at what point of degradation they would like to try a restart.