feat: correctly deal with advanced synchronizers
Previously if all synchronizer would be advanced we would unnecessarily crash the block stream.

Now if any synchronizer is advanced, we use that block as the new latest block and update the remaining synchronizer states accordingly.

This way the advanced state now is purely internal and never exposed to clients.