feat: Implement fluid delta_transition
This function is custom since it calls the VM to rebuild the full state, then update the internal state from that.

To enable this to work correctly and catch all state changes, we need to call this method also if a DCI indexed contract had a change. Therefore the contracts_map in the Decoder was changed accordingly as well.