fix(EVMPoolState): After swapping, set the spot prices correctly
Using the inverse of the price of the trade is wrong because of the fee. The trade price already has the fee deducted in this direction, to get the correct price of the inverse direction we need use the non fee price - which we don't have access to.
Using set_spot_prices will deliver correct results, but it will make get_amount_out a little bit slower

Took 25 minutes


Took 2 minutes